Separando gráficos da implementação
As funções de gréficos está separada da camada física. Projeto com alguns comentários adicionados
