+++ CamadaFisica.hpp
@@ -12,30 +12,23 @@ using namespace std;
 
 string converter_mensagem_em_bit(string mensagem);
 
-// Construir Sinal
-
-void construir_sinal(vector<char> quadro);
-void construir_sinal_binario(vector<char> quadro);
-void construir_sinal_bipolar(vector<char> quadro);
-void construir_sinal_manchester(vector<char> quadro);
-
 // Transmissão
 
-void aplicacao_transmissora();
-void camada_de_aplicacao_transmissora(string mensagem);
-void camada_fisica_transmissora(vector<char> quadro);
+string aplicacao_transmissora(void);
+vector<char> camada_de_aplicacao_transmissora(string mensagem);
+vector<char> camada_fisica_transmissora(vector<char> quadro, int tipo_de_codificacao);
 vector<char> camada_fisica_transmissora_codificacao_bipolar(vector<char> quadro);
 vector<char> camada_fisica_transmissora_codificacao_binaria(vector<char> quadro);
 vector<char> camada_fisica_transmissora_codificacao_manchester(vector<char> quadro);
 
 // Comunicação
 
-void meio_de_comunicacao(vector<char> fluxo_bruto_de_bits);
+vector<char> meio_de_comunicacao(vector<char> fluxo_bruto_de_bits);
 
 // Recepçao
 
-void camada_fisica_receptora(vector<char> quadro);
-void camada_de_aplicacao_receptora(vector<char> quadro);
+vector<char> camada_fisica_receptora(vector<char> quadro, int tipo_de_codificacao);
+string camada_de_aplicacao_receptora(vector<char> quadro);
 vector<char> camada_fisica_receptora_decodificacao_bipolar(vector<char> quadro);
 vector<char> camada_fisica_receptora_decodificacao_binaria(vector<char> quadro);
 vector<char> camada_fisica_receptora_decodificacao_manchester(vector<char> quadro);
