+++ makefile
@@ -1,0 +1,9 @@
+cpp_compile:
+		g++ ./*.cpp -o executar
+
+cpp_run:
+		./executar
+
+cpp_cr:
+		g++ ./*.cpp -o executar
+		./executar
